+++ tests/IntegrationTestCase.php
@@ -2,13 +2,15 @@
 
 declare(strict_types=1);
 
-namespace NotificationChannels\Twilio\Tests\Integration;
+namespace NotificationChannels\Twilio\Tests;
 
 use NotificationChannels\Twilio\TwilioProvider;
 use Orchestra\Testbench\TestCase;
 
-abstract class BaseIntegrationTest extends TestCase
+class IntegrationTestCase extends TestCase
 {
+    protected static $latestResponse;
+
     protected function getPackageProviders($app)
     {
         return [TwilioProvider::class];