+++ tests/Unit/IntegrationTest.php
@@ -3,14 +3,17 @@
 namespace NotificationChannels\Twilio\Tests\Unit;
 
 use Illuminate\Contracts\Events\Dispatcher;
+use Illuminate\Notifications\Events\NotificationFailed;
 use Illuminate\Notifications\Notification;
 use Mockery;
 use Mockery\Adapter\Phpunit\MockeryTestCase;
+use NotificationChannels\Twilio\Exceptions\CouldNotSendNotification;
 use NotificationChannels\Twilio\Twilio;
 use NotificationChannels\Twilio\TwilioCallMessage;
 use NotificationChannels\Twilio\TwilioChannel;
 use NotificationChannels\Twilio\TwilioConfig;
 use NotificationChannels\Twilio\TwilioSmsMessage;
+use PHPUnit\Framework\Attributes\Test;
 use Twilio\Rest\Api\V2010\Account\CallInstance;
 use Twilio\Rest\Api\V2010\Account\CallList;
 use Twilio\Rest\Api\V2010\Account\MessageInstance;
@@ -28,7 +31,7 @@ class IntegrationTest extends MockeryTestCase
     /** @var Dispatcher */
     protected $events;
 
-    public function setUp(): void
+    protected function setUp(): void
     {
         parent::setUp();
 
@@ -40,7 +43,7 @@ class IntegrationTest extends MockeryTestCase
         $this->notification = Mockery::mock(Notification::class);
     }
 
-    /** @test */
+    #[Test]
     public function it_can_send_a_sms_message()
     {
         $message = TwilioSmsMessage::create('Message text');
@@ -57,10 +60,10 @@ class IntegrationTest extends MockeryTestCase
             'body' => 'Message text',
         ]);
 
-        $channel->send(new NotifiableWithAttribute(), $this->notification);
+        $channel->send(new NotifiableWithAttribute, $this->notification);
     }
 
-    /** @test */
+    #[Test]
     public function it_can_send_a_sms_message_using_service()
     {
         $message = TwilioSmsMessage::create('Message text');
@@ -79,10 +82,10 @@ class IntegrationTest extends MockeryTestCase
             'messagingServiceSid' => '0123456789',
         ]);
 
-        $channel->send(new NotifiableWithAttribute(), $this->notification);
+        $channel->send(new NotifiableWithAttribute, $this->notification);
     }
 
-    /** @test */
+    #[Test]
     public function it_can_send_a_sms_message_using_url_shortener()
     {
         $message = TwilioSmsMessage::create('Message text');
@@ -90,7 +93,7 @@ class IntegrationTest extends MockeryTestCase
 
         $config = new TwilioConfig([
             'from' => '+31612345678',
-            'ShortenUrls' => true,
+            'shorten_urls' => true,
         ]);
         $twilio = new Twilio($this->twilioService, $config);
         $channel = new TwilioChannel($twilio, $this->events);
@@ -101,10 +104,10 @@ class IntegrationTest extends MockeryTestCase
             'ShortenUrls' => 'true',
         ]);
 
-        $channel->send(new NotifiableWithAttribute(), $this->notification);
+        $channel->send(new NotifiableWithAttribute, $this->notification);
     }
 
-    /** @test */
+    #[Test]
     public function it_can_send_a_sms_message_using_alphanumeric_sender()
     {
         $message = TwilioSmsMessage::create('Message text');
@@ -122,10 +125,10 @@ class IntegrationTest extends MockeryTestCase
             'body' => 'Message text',
         ]);
 
-        $channel->send(new NotifiableWithAlphanumericSender(), $this->notification);
+        $channel->send(new NotifiableWithAlphanumericSender, $this->notification);
     }
 
-    /** @test */
+    #[Test]
     public function it_can_make_a_call()
     {
         $message = TwilioCallMessage::create('http://example.com');
@@ -141,7 +144,28 @@ class IntegrationTest extends MockeryTestCase
             'url' => 'http://example.com',
         ]);
 
-        $channel->send(new NotifiableWithAttribute(), $this->notification);
+        $channel->send(new NotifiableWithAttribute, $this->notification);
+    }
+
+    #[Test]
+    public function it_cant_make_a_call_when_the_from_config_is_missing()
+    {
+        $message = TwilioCallMessage::create('http://example.com');
+        $this->notification->shouldReceive('toTwilio')->andReturn($message);
+
+        $config = new TwilioConfig([]);
+        $twilio = new Twilio($this->twilioService, $config);
+        $channel = new TwilioChannel($twilio, $this->events);
+
+        $this->twilioService->calls->shouldNotReceive('create');
+
+        $this->events->shouldReceive('dispatch')
+            ->atLeast()->once()
+            ->with(Mockery::type(NotificationFailed::class));
+
+        $this->expectException(CouldNotSendNotification::class);
+
+        $channel->send(new NotifiableWithAttribute, $this->notification);
     }
 
     protected function smsMessageWillBeSentToTwilioWith(...$args)