+++ tests/Unit/TwilioChannelTest.php
@@ -13,6 +13,7 @@ use NotificationChannels\Twilio\TwilioCallMessage;
 use NotificationChannels\Twilio\TwilioChannel;
 use NotificationChannels\Twilio\TwilioConfig;
 use NotificationChannels\Twilio\TwilioSmsMessage;
+use PHPUnit\Framework\Attributes\Test;
 use Twilio\Exceptions\RestException;
 
 class TwilioChannelTest extends MockeryTestCase
@@ -26,20 +27,38 @@ class TwilioChannelTest extends MockeryTestCase
     /** @var Dispatcher */
     protected $dispatcher;
 
-    public function setUp(): void
+    protected function setUp(): void
     {
         parent::setUp();
 
         $this->twilio = Mockery::mock(Twilio::class);
+        $this->twilio->config = new TwilioConfig([]);
         $this->dispatcher = Mockery::mock(Dispatcher::class);
 
         $this->channel = new TwilioChannel($this->twilio, $this->dispatcher);
     }
 
-    /** @test */
+    #[Test]
+    public function it_will_not_send_a_message_if_not_enabled()
+    {
+        $notifiable = new Notifiable;
+        $notification = Mockery::mock(Notification::class);
+
+        $this->twilio->config = new TwilioConfig([
+            'enabled' => false,
+        ]);
+
+        $this->dispatcher->shouldNotReceive('dispatch');
+
+        $result = $this->channel->send($notifiable, $notification);
+
+        $this->assertNull($result);
+    }
+
+    #[Test]
     public function it_will_not_send_a_message_without_known_receiver()
     {
-        $notifiable = new Notifiable();
+        $notifiable = new Notifiable;
         $notification = Mockery::mock(Notification::class);
 
         $this->twilio->config = new TwilioConfig([
@@ -57,10 +76,10 @@ class TwilioChannelTest extends MockeryTestCase
         $this->assertNull($result);
     }
 
-    /** @test */
+    #[Test]
     public function it_will_send_a_sms_message_to_the_result_of_the_route_method_of_the_notifiable()
     {
-        $notifiable = new NotifiableWithMethod();
+        $notifiable = new NotifiableWithMethod;
         $notification = Mockery::mock(Notification::class);
 
         $message = new TwilioSmsMessage('Message text');
@@ -73,10 +92,26 @@ class TwilioChannelTest extends MockeryTestCase
         $this->channel->send($notifiable, $notification);
     }
 
-    /** @test */
+    #[Test]
+    public function it_will_send_a_sms_message_to_the_result_of_the_route_method_of_the_notifiable_if_it_uses_the_twilio_channel_explicitly()
+    {
+        $notifiable = new NotifiableWithTwilioChannel;
+        $notification = Mockery::mock(Notification::class);
+
+        $message = new TwilioSmsMessage('Message text');
+        $notification->shouldReceive('toTwilio')->andReturn($message);
+
+        $this->twilio->shouldReceive('sendMessage')
+            ->atLeast()->once()
+            ->with($message, '+1111111111', false);
+
+        $this->channel->send($notifiable, $notification);
+    }
+
+    #[Test]
     public function it_will_make_a_call_to_the_phone_number_attribute_of_the_notifiable()
     {
-        $notifiable = new NotifiableWithAttribute();
+        $notifiable = new NotifiableWithAttribute;
         $notification = Mockery::mock(Notification::class);
 
         $message = new TwilioCallMessage('http://example.com');
@@ -89,10 +124,10 @@ class TwilioChannelTest extends MockeryTestCase
         $this->channel->send($notifiable, $notification);
     }
 
-    /** @test */
+    #[Test]
     public function it_will_convert_a_string_to_a_sms_message()
     {
-        $notifiable = new NotifiableWithAttribute();
+        $notifiable = new NotifiableWithAttribute;
         $notification = Mockery::mock(Notification::class);
 
         $notification->shouldReceive('toTwilio')->andReturn('Message text');
@@ -104,10 +139,10 @@ class TwilioChannelTest extends MockeryTestCase
         $this->channel->send($notifiable, $notification);
     }
 
-    /** @test */
+    #[Test]
     public function it_will_fire_an_event_in_case_of_an_invalid_message()
     {
-        $notifiable = new NotifiableWithAttribute();
+        $notifiable = new NotifiableWithAttribute;
         $notification = Mockery::mock(Notification::class);
 
         $this->twilio->config = new TwilioConfig([
@@ -126,10 +161,10 @@ class TwilioChannelTest extends MockeryTestCase
         $this->channel->send($notifiable, $notification);
     }
 
-    /** @test */
+    #[Test]
     public function it_will_ignore_specific_error_codes()
     {
-        $notifiable = new NotifiableWithAttribute();
+        $notifiable = new NotifiableWithAttribute;
         $notification = Mockery::mock(Notification::class);
 
         $this->twilio->config = new TwilioConfig([
@@ -150,10 +185,10 @@ class TwilioChannelTest extends MockeryTestCase
         $this->channel->send($notifiable, $notification);
     }
 
-    /** @test */
+    #[Test]
     public function it_will_rethrow_non_ignored_error_codes()
     {
-        $notifiable = new NotifiableWithAttribute();
+        $notifiable = new NotifiableWithAttribute;
         $notification = Mockery::mock(Notification::class);
 
         $this->twilio->config = new TwilioConfig([
@@ -176,10 +211,10 @@ class TwilioChannelTest extends MockeryTestCase
         $this->channel->send($notifiable, $notification);
     }
 
-    /** @test */
+    #[Test]
     public function it_will_ignore_all_error_codes()
     {
-        $notifiable = new NotifiableWithAttribute();
+        $notifiable = new NotifiableWithAttribute;
         $notification = Mockery::mock(Notification::class);
 
         $this->twilio->config = new TwilioConfig([
@@ -198,10 +233,10 @@ class TwilioChannelTest extends MockeryTestCase
         $this->channel->send($notifiable, $notification);
     }
 
-    /** @test */
+    #[Test]
     public function it_will_send_using_alphanumeric_if_notifiable_can_receive()
     {
-        $notifiable = new NotifiableWithAlphanumericSender();
+        $notifiable = new NotifiableWithAlphanumericSender;
         $notification = Mockery::mock(Notification::class);
 
         $message = new TwilioSmsMessage('Message text');
@@ -219,8 +254,16 @@ class Notifiable
 {
     public $phone_number = null;
 
-    public function routeNotificationFor()
+    public function routeNotificationFor() {}
+}
+
+class NotifiableWithTwilioChannel
+{
+    public function routeNotificationFor(string $channel)
     {
+        if ($channel === 'twilio') {
+            return '+1111111111';
+        }
     }
 }
 
@@ -236,18 +279,14 @@ class NotifiableWithAttribute
 {
     public $phone_number = '+22222222222';
 
-    public function routeNotificationFor()
-    {
-    }
+    public function routeNotificationFor() {}
 }
 
 class NotifiableWithAlphanumericSender
 {
     public $phone_number = '+33333333333';
 
-    public function routeNotificationFor()
-    {
-    }
+    public function routeNotificationFor() {}
 
     public function canReceiveAlphanumericSender()
     {