+++ tests/Unit/TwilioTest.php
@@ -12,6 +12,7 @@ use NotificationChannels\Twilio\TwilioConfig;
 use NotificationChannels\Twilio\TwilioMessage;
 use NotificationChannels\Twilio\TwilioMmsMessage;
 use NotificationChannels\Twilio\TwilioSmsMessage;
+use PHPUnit\Framework\Attributes\Test;
 use Twilio\Rest\Api\V2010\Account\CallInstance;
 use Twilio\Rest\Api\V2010\Account\CallList;
 use Twilio\Rest\Api\V2010\Account\MessageInstance;
@@ -34,7 +35,7 @@ class TwilioTest extends MockeryTestCase
      */
     protected $config;
 
-    public function setUp(): void
+    protected function setUp(): void
     {
         parent::setUp();
 
@@ -48,7 +49,7 @@ class TwilioTest extends MockeryTestCase
         $this->twilio = new Twilio($this->twilioService, $this->config);
     }
 
-    /** @test */
+    #[Test]
     public function it_can_send_a_sms_message_to_twilio()
     {
         $message = new TwilioSmsMessage('Message text');
@@ -66,6 +67,10 @@ class TwilioTest extends MockeryTestCase
         $this->config->shouldReceive('getDebugTo')
             ->once()
             ->andReturn(null);
+
+        $this->config->shouldReceive('isShortenUrlsEnabled')
+            ->once()
+            ->andReturn(false);
 
         $this->config->shouldReceive('getServiceSid')
             ->once()
@@ -88,7 +93,7 @@ class TwilioTest extends MockeryTestCase
         $this->twilio->sendMessage($message, '+1111111111');
     }
 
-    /** @test */
+    #[Test]
     public function it_can_send_a_mms_message_to_twilio()
     {
         $message = new TwilioMmsMessage('Message text');
@@ -103,6 +108,10 @@ class TwilioTest extends MockeryTestCase
         $this->config->shouldReceive('getFrom')
             ->once()
             ->andReturn('+1234567890');
+
+        $this->config->shouldReceive('isShortenUrlsEnabled')
+            ->once()
+            ->andReturn(false);
 
         $this->config->shouldReceive('getServiceSid')
             ->once()
@@ -130,7 +139,7 @@ class TwilioTest extends MockeryTestCase
         $this->twilio->sendMessage($message, '+1111111111');
     }
 
-    /** @test */
+    #[Test]
     public function it_can_send_a_sms_message_to_twilio_with_alphanumeric_sender()
     {
         $message = new TwilioSmsMessage('Message text');
@@ -140,6 +149,10 @@ class TwilioTest extends MockeryTestCase
             ->andReturn('TwilioTest');
 
         $this->config->shouldNotReceive('getFrom');
+
+        $this->config->shouldReceive('isShortenUrlsEnabled')
+            ->once()
+            ->andReturn(false);
 
         $this->config->shouldReceive('getServiceSid')
             ->once()
@@ -160,7 +173,7 @@ class TwilioTest extends MockeryTestCase
         $this->twilio->sendMessage($message, '+1111111111', true);
     }
 
-    /** @test */
+    #[Test]
     public function it_can_send_a_sms_message_to_twilio_with_messaging_service()
     {
         $message = new TwilioSmsMessage('Message text');
@@ -168,6 +181,10 @@ class TwilioTest extends MockeryTestCase
         $this->config->shouldReceive('getFrom')
             ->once()
             ->andReturn('+1234567890');
+
+        $this->config->shouldReceive('isShortenUrlsEnabled')
+            ->once()
+            ->andReturn(false);
 
         $this->config->shouldReceive('getServiceSid')
             ->once()
@@ -189,7 +206,7 @@ class TwilioTest extends MockeryTestCase
         $this->twilio->sendMessage($message, '+1111111111');
     }
 
-    /** @test */
+    #[Test]
     public function it_can_send_a_call_to_twilio()
     {
         $message = new TwilioCallMessage('http://example.com');
@@ -200,6 +217,10 @@ class TwilioTest extends MockeryTestCase
         $message->statusCallbackMethod('PUT');
         $message->fallbackUrl('http://example.com');
         $message->fallbackMethod('PUT');
+
+        $this->config->shouldReceive('getDebugTo')
+            ->once()
+            ->andReturn('+1111111111');
 
         $this->twilioService->calls->shouldReceive('create')
             ->atLeast()->once()
@@ -217,7 +238,7 @@ class TwilioTest extends MockeryTestCase
         $this->twilio->sendMessage($message, '+1111111111');
     }
 
-    /** @test */
+    #[Test]
     public function it_will_throw_an_exception_in_case_of_a_missing_from_number()
     {
         $this->expectException(CouldNotSendNotification::class);
@@ -233,6 +254,10 @@ class TwilioTest extends MockeryTestCase
             ->once()
             ->andReturn(null);
 
+        $this->config->shouldReceive('isShortenUrlsEnabled')
+            ->once()
+            ->andReturn(false);
+
         $this->config->shouldReceive('getServiceSid')
             ->once()
             ->andReturn(null);
@@ -240,16 +265,16 @@ class TwilioTest extends MockeryTestCase
         $this->twilio->sendMessage($smsMessage, null);
     }
 
-    /** @test */
+    #[Test]
     public function it_will_throw_an_exception_in_case_of_an_unrecognized_message_object()
     {
         $this->expectException(CouldNotSendNotification::class);
         $this->expectExceptionMessage('Notification was not sent. Message object class');
 
-        $this->twilio->sendMessage(new InvalidMessage(), null);
+        $this->twilio->sendMessage(new InvalidMessage, null);
     }
 
-    /** @test */
+    #[Test]
     public function it_should_use_universal_to()
     {
         $debugTo = '+1222222222';
@@ -269,6 +294,10 @@ class TwilioTest extends MockeryTestCase
         $this->config->shouldReceive('getDebugTo')
             ->once()
             ->andReturn($debugTo);
+
+        $this->config->shouldReceive('isShortenUrlsEnabled')
+            ->once()
+            ->andReturn(false);
 
         $this->config->shouldReceive('getServiceSid')
             ->once()
@@ -292,6 +321,4 @@ class TwilioTest extends MockeryTestCase
     }
 }
 
-class InvalidMessage extends TwilioMessage
-{
-}
+class InvalidMessage extends TwilioMessage {}